+++ cli/commands/list_command.h
@@ -7,9 +7,10 @@ class ListCommand : public BaseCommand {
 public:
     ListCommand(const WorkspaceServicePtr& service);
 
-    void reg(Args::CmdLine& cmdLine) override;
-    bool process(const Args::CmdLine& cmdLine) override;
+    void reg(CLI::App& app) override;
 
 private:
+    void process();
     WorkspaceServicePtr service;
+    bool namesOnly;
 };