+++ cli/commands/up_command.h
@@ -7,9 +7,11 @@ class UpCommand : public BaseCommand {
 public:
     UpCommand(const WorkspaceServicePtr& service);
 
-    void reg(Args::CmdLine& cmdLine) override;
-    bool process(const Args::CmdLine& cmdLine) override;
+    void reg(CLI::App& app) override;
 
 private:
+    void process();
     WorkspaceServicePtr service;
+    std::string name;
+    bool clean;
 };