+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.0 (2024.05.08)
+
+* Switch to new command line parser [CLI11](https://github.com/CLIUtils/CLI11)
+* Suggest workspace name in `up` command if it not provided. Name suggested if current directory matches workspace or 
+  try to up the latest active workspace.
+* Added shortcuts for all commands
+* `dcw` without args shows a list of all workspaces
+
 # 1.1.2 (2024.04.24)
 
 * Fix location of state yaml file to `~/.local/share/dcw/state.yml`